refactor and tidy-up sidecar.
* remove the `dockermanager` package; merge into `docker`.
* fix overusage of term "manager"/"manage".
* shard large source files.
