+++ docs/DEEP_TRACING.md
@@ -1,0 +1,11 @@
+## Design blueprint: Deep tracing
+
+DHT PR for automatic client/server mode switching. Things we want to measure:
+
+1. hop count to find a server
+2. hop count to find a client
+3. \# peers queried at each distance
+4. time between distance hops
+5. 
+
+and we want
